+++ src/main/java/com/mev/films/service/implement/DiscountServiceImpl.java
@@ -39,7 +39,16 @@ public class DiscountServiceImpl implements DiscountService{
         LOG.debug("getDiscount: id = {}",
                 id);
 
-        return discountMapper.selectDiscount(id);
+        DiscountDTO discountDTO = null;
+        if (id != null && id >= 0){
+            discountDTO = discountMapper.selectDiscount(id);
+        }
+        else {
+            LOG.debug("Error in 'getDiscount'! 'id' is not validate: id = {}",
+                    id);
+        }
+
+        return discountDTO;
     }
 
     @Override
@@ -47,7 +56,15 @@ public class DiscountServiceImpl implements DiscountService{
         LOG.debug("getDiscountByCode: code = {}",
                 code);
 
-        return discountMapper.selectDiscountByCode(code);
+        DiscountDTO discountDTO = null;
+        if (code != null){
+            discountDTO = discountMapper.selectDiscountByCode(code);
+        }
+        else {
+            LOG.debug("Error on 'getDiscountByCode'! 'code' is null");
+        }
+
+        return discountDTO;
     }
 
     @Override
@@ -55,7 +72,25 @@ public class DiscountServiceImpl implements DiscountService{
         LOG.debug("addDiscount: discountDTO = {}",
                 discountDTO);
 
-        discountMapper.insertDiscount(discountDTO);
+        if (discountDTO != null){
+            if (discountDTO.getCode() != null){
+                if (discountDTO.getValue() > 0 && discountDTO.getValue() < 0.75F){
+
+                    discountMapper.insertDiscount(discountDTO);
+
+                }
+                else {
+                    LOG.debug("Error in 'addDiscount!' 'value' is not validate: value = {}",
+                            discountDTO.getValue());
+                }
+            }
+            else {
+                LOG.debug("Error in 'addDiscount!' 'code' is null");
+            }
+        }
+        else {
+            LOG.debug("Error in 'addDiscount!' 'discountDTO' is null");
+        }
     }
 
     @Override
@@ -63,7 +98,45 @@ public class DiscountServiceImpl implements DiscountService{
         LOG.debug("updateDiscount: discountDTO = {}",
                 discountDTO);
 
+        if (discountDTO != null) {
+            if (discountDTO.getId() != null && discountDTO.getId() >= 0) {
+                if (discountDTO.getCode() != null) {
+                    if (discountDTO.getValue() > 0 && discountDTO.getValue() < 0.75F) {
+
+                        discountMapper.updateDiscount(discountDTO);
+
+                    } else {
+                        LOG.debug("Error in 'updateDiscount!' 'value' is not validate: value = {}",
+                                discountDTO.getValue());
+                    }
+                } else {
+                    LOG.debug("Error in 'updateDiscount!' 'code' is null");
+                }
+            }
+            else {
+                LOG.debug("Error in 'updateDiscount'! 'id' is not validate: id = {}",
+                        discountDTO.getId());
+            }
+        }
+        else {
+            LOG.debug("Error in 'updateDiscount!' 'discountDTO' is null");
+        }
+
         discountMapper.updateDiscount(discountDTO);
+    }
+
+    @Override
+    public void deleteDiscount(Long id) {
+        LOG.debug("deleteDiscount: id = {}",
+                id);
+
+        if (id != null && id >= 0){
+            discountMapper.deleteDiscount(id);
+        }
+        else {
+            LOG.debug("Error in 'deleteDiscount'! 'id' is not validate: id = {}",
+                    id);
+        }
     }
 
     @Override
@@ -71,6 +144,11 @@ public class DiscountServiceImpl implements DiscountService{
         LOG.debug("deleteDiscountByCode: code = {}",
                 code);
 
-        discountMapper.deleteDiscountByCode(code);
+        if (code != null){
+            discountMapper.deleteDiscountByCode(code);
+        }
+        else {
+            LOG.debug("Error in 'deleteDiscountByCode'! 'code' is null");
+        }
     }
 }
