+++ src/main/java/com/mev/films/service/implement/FilmServiceImpl.java
@@ -40,7 +40,16 @@ public class FilmServiceImpl implements FilmService{
         LOG.debug("getFilm: id = {}",
                 id);
 
-        return filmMapper.selectFilm(id);
+        FilmDTO filmDTO = null;
+        if (id != null && id >= 0){
+            filmDTO = filmMapper.selectFilm(id);
+        }
+        else {
+            LOG.debug("Error in 'getFilm'! 'id' is not validate: id = {}",
+                    id);
+        }
+
+        return filmDTO;
     }
 
     @Override
@@ -55,7 +64,15 @@ public class FilmServiceImpl implements FilmService{
         LOG.debug("getFilmByName: name = {}",
                 name);
 
-        return filmMapper.selectFilmsByName(name);
+        List<FilmDTO> filmDTOS = null;
+        if (name != null){
+            filmDTOS = filmMapper.selectFilmsSortByName();
+        }
+        else {
+            LOG.debug("Error in 'getFilmByName'! 'name' is null");
+        }
+
+        return filmDTOS;
     }
 
     @Override
@@ -63,14 +80,66 @@ public class FilmServiceImpl implements FilmService{
         LOG.debug("getFilmByImage: image = {}",
                 image);
 
-        return filmMapper.selectFilmByImage(image);
+        FilmDTO filmDTO = null;
+        if (image != null && (image.endsWith(".jpg") ||
+                              image.endsWith(".jpeg") ||
+                              image.endsWith(".png") ||
+                              image.endsWith(".bmp"))){
+            filmDTO = filmMapper.selectFilmByImage(image);
+        }
+        else {
+            LOG.debug("Error in 'getFilmByImage'! image is not validate: image = {}",
+                    image);
+        }
+
+        return filmDTO;
     }
 
     @Override
     public void addFilm(FilmDTO filmDTO) {
-        LOG.debug("addFilm: filmDTO = {}",
+        LOG.debug("addFilm: {}",
                 filmDTO);
-        filmMapper.insertFilm(filmDTO);
+
+        if (filmDTO != null){
+            if (filmDTO.getName() != null){
+                if (filmDTO.getGenre() != null){
+                    if (filmDTO.getPrice() > 0 && filmDTO.getPrice() < 10000){
+                        if (filmDTO.getDuration() > 0 && filmDTO.getDuration() < 300) {
+                            String image = filmDTO.getImage();
+                            if (image != null && (image.endsWith(".jpg") ||
+                                    image.endsWith(".jpeg") ||
+                                    image.endsWith(".png") ||
+                                    image.endsWith(".bmp"))){
+
+                                filmMapper.insertFilm(filmDTO);
+
+                            }
+                            else {
+                                LOG.debug("Error in 'addFilm'! 'image' is not validate: image = {}",
+                                        filmDTO.getImage());
+                            }
+                        }
+                        else {
+                            LOG.debug("Error in 'addFilm'! 'duration' is not validate: duration = {}",
+                                    filmDTO.getDuration());
+                        }
+                    }
+                    else {
+                        LOG.debug("Error in 'addFilm'! 'price' is not validate: price = {}",
+                                filmDTO.getPrice());
+                    }
+                }
+                else {
+                    LOG.debug("Error in 'addFilm'! 'genre' is null");
+                }
+            }
+            else {
+                LOG.debug("Error in 'addFilm'! 'name' is null");
+            }
+        }
+        else {
+            LOG.debug("Error in 'addFilm'! 'filmDTO' is null");
+        }
     }
 
     @Override
@@ -78,7 +147,62 @@ public class FilmServiceImpl implements FilmService{
         LOG.debug("updateFilm: filmDTO = {}",
                 filmDTO);
 
-        filmMapper.updateFilm(filmDTO);
+        if (filmDTO != null){
+            if (filmDTO.getId() != null && filmDTO.getId() >= 0) {
+                if (filmDTO.getName() != null) {
+                    if (filmDTO.getGenre() != null) {
+                        if (filmDTO.getPrice() > 0 && filmDTO.getPrice() < 10000) {
+                            if (filmDTO.getDuration() > 0 && filmDTO.getDuration() < 300) {
+                                String image = filmDTO.getImage();
+                                if (image != null && (image.endsWith(".jpg") ||
+                                        image.endsWith(".jpeg") ||
+                                        image.endsWith(".png") ||
+                                        image.endsWith(".bmp"))) {
+
+                                    filmMapper.updateFilm(filmDTO);
+
+                                } else {
+                                    LOG.debug("Error in 'addFilm'! 'image' is not validate: image = {}",
+                                            image);
+                                }
+                            } else {
+                                LOG.debug("Error in 'addFilm'! 'duration' is not validate: duration = {}",
+                                        filmDTO.getDuration());
+                            }
+                        } else {
+                            LOG.debug("Error in 'addFilm'! 'price' is not validate: price = {}",
+                                    filmDTO.getPrice());
+                        }
+                    } else {
+                        LOG.debug("Error in 'addFilm'! 'genre' is null");
+                    }
+                } else {
+                    LOG.debug("Error in 'addFilm'! 'name' is null");
+                }
+            }
+            else {
+                LOG.debug("Error in 'addFilm'! 'id' is not validate: id = {}",
+                        filmDTO.getId());
+            }
+        }
+        else {
+            LOG.debug("Error in 'addFilm'! 'filmDTO' is null");
+        }
+    }
+
+    @Override
+    public void deleteFilm(Long id) {
+
+        LOG.debug("deleteFilm: id = {}",
+                id);
+
+        if (id != null && id >= 0){
+            filmMapper.deleteFilm(id);
+        }
+        else {
+            LOG.debug("Error in 'deleteFilm'! 'id' is not validate: id = {}",
+                    id);
+        }
     }
 
     @Override
@@ -86,6 +210,15 @@ public class FilmServiceImpl implements FilmService{
         LOG.debug("deleteFilmByImage: image = {}",
                 image);
 
-        filmMapper.deleteFilmByImage(image);
+        if (image != null && (image.endsWith(".jpg") ||
+                image.endsWith(".jpeg") ||
+                image.endsWith(".png") ||
+                image.endsWith(".bmp"))) {
+
+            filmMapper.deleteFilmByImage(image);
+        } else {
+            LOG.debug("Error in 'addFilm'! 'image' is not validate: image = {}",
+                    image);
+        }
     }
 }
