+++ src/main/java/com/mev/films/service/implement/UserDiscountServiceImpl.java
@@ -1,8 +1,12 @@
 package com.mev.films.service.implement;
 
 import com.mev.films.mappers.interfaces.UserDiscountMapper;
+import com.mev.films.model.DiscountDTO;
+import com.mev.films.model.UserDTO;
 import com.mev.films.model.UserDiscountDTO;
+import com.mev.films.service.interfaces.DiscountService;
 import com.mev.films.service.interfaces.UserDiscountService;
+import com.mev.films.service.interfaces.UserService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,12 +23,66 @@ public class UserDiscountServiceImpl implements UserDiscountService{
 
     @Autowired private UserDiscountMapper userDiscountMapper;
 
-    public UserDiscountServiceImpl(){
+    @Autowired private UserService userService;
+    @Autowired private DiscountService discountService;
 
+    public UserDiscountServiceImpl(){
     }
 
     public UserDiscountServiceImpl(UserDiscountMapper userDiscountMapper){
         this.userDiscountMapper = userDiscountMapper;
+    }
+
+    private boolean checkUserDiscount(UserDiscountDTO userDiscountDTO, boolean idCheck, String functionName){
+        boolean userOk = false;
+        boolean discountOk = false;
+        boolean idOk = true;
+
+        UserDTO userDTO = userDiscountDTO.getUserDTO();
+        if (userDTO != null){
+            Long userId = userDTO.getId();
+            if (userId != null && userId  >= 0){
+                if (userService.getUser(userId) != null){
+                    userOk = true;
+                }
+                else {
+                    LOG.debug("Error in " + "'" + functionName + "'! User not found");
+                }
+            }
+            else {
+                LOG.debug("Error in " + "'" + functionName + "'! 'user_id is not validate: user_id = {}",
+                        userId);
+            }
+        }
+        else {
+            LOG.debug("Error in " + "'" + functionName + "'! 'userDTO' is null");
+        }
+
+        DiscountDTO discountDTO = userDiscountDTO.getDiscountDTO();
+        if (discountDTO != null){
+            Long discountId = discountDTO.getId();
+            if (discountId != null && discountId >= 0){
+                if (discountService.getDiscount(discountId) != null){
+                    discountOk = true;
+                }
+                else {
+                    LOG.debug("Error in " + "'" + functionName + "'! Discount not found");
+                }
+            }
+            else {
+                LOG.debug("Error in " + "'" + functionName + "'! 'discount_id' is not validate: discount_id = {}",
+                        discountId);
+            }
+        }
+        else {
+            LOG.debug("Error in " + "'" + functionName + "'! 'discountDTO' is null");
+        }
+
+        if (idCheck){
+            idOk = userDiscountDTO.getId() != null && userDiscountDTO.getId() >= 0;
+        }
+
+        return userOk && discountOk && idOk;
     }
 
     @Override
@@ -39,7 +97,16 @@ public class UserDiscountServiceImpl implements UserDiscountService{
         LOG.debug("getUserDiscount: id = {}",
                 id);
 
-        return userDiscountMapper.selectUserDiscount(id);
+        UserDiscountDTO userDiscountDTO = null;
+        if (id != null && id >= 0){
+            userDiscountDTO = userDiscountMapper.selectUserDiscount(id);
+        }
+        else {
+            LOG.debug("Error in 'getUserDiscount'! 'id' is not validate: id = {}",
+                    id);
+        }
+
+        return userDiscountDTO;
     }
 
     @Override
@@ -47,7 +114,16 @@ public class UserDiscountServiceImpl implements UserDiscountService{
         LOG.debug("getUserDiscountsByUser: user_id = {}",
                 userId);
 
-        return userDiscountMapper.selectUserDiscountsByUser(userId);
+        List<UserDiscountDTO> userDiscountDTOS = null;
+        if (userId != null && userId >= 0){
+            userDiscountDTOS = userDiscountMapper.selectUserDiscountsByUser(userId);
+        }
+        else {
+            LOG.debug("Error in 'getUserDiscountsByUser'! 'user_id is not validate: user_id = {}",
+                    userId);
+        }
+
+        return userDiscountDTOS;
     }
 
     @Override
@@ -55,7 +131,16 @@ public class UserDiscountServiceImpl implements UserDiscountService{
         LOG.debug("getUserDiscountByDiscount: discount_id = {}",
                 discountId);
 
-        return userDiscountMapper.selectUserDiscountByDiscount(discountId);
+        UserDiscountDTO userDiscountDTO = null;
+        if (discountId != null && discountId >= 0){
+            userDiscountDTO = userDiscountMapper.selectUserDiscountByDiscount(discountId);
+        }
+        else {
+            LOG.debug("Error in 'getUserDiscountByDiscount'! 'discount_id is not validate: discount_id = {}",
+                    discountId);
+        }
+
+        return userDiscountDTO;
     }
 
     @Override
@@ -63,7 +148,9 @@ public class UserDiscountServiceImpl implements UserDiscountService{
         LOG.debug("addUserDiscount: userDiscountDTO = {}",
                 userDiscountDTO);
 
-        userDiscountMapper.insertUserDiscount(userDiscountDTO);
+        if (checkUserDiscount(userDiscountDTO, false, "addUserDiscount")) {
+            userDiscountMapper.insertUserDiscount(userDiscountDTO);
+        }
     }
 
     @Override
@@ -71,7 +158,9 @@ public class UserDiscountServiceImpl implements UserDiscountService{
         LOG.debug("updateUserDiscount: userDiscountDTO = {}",
                 userDiscountDTO);
 
-        userDiscountMapper.updateUserDiscount(userDiscountDTO);
+        if (checkUserDiscount(userDiscountDTO, true, "updateUserDiscount")) {
+            userDiscountMapper.updateUserDiscount(userDiscountDTO);
+        }
     }
 
     @Override
@@ -79,7 +168,13 @@ public class UserDiscountServiceImpl implements UserDiscountService{
         LOG.debug("deleteUserDiscount: id = {}",
                 id);
 
-        userDiscountMapper.deleteUserDiscount(id);
+        if (id != null && id >= 0){
+            userDiscountMapper.deleteUserDiscount(id);
+        }
+        else {
+            LOG.debug("Error in 'deleteUserDiscount'! 'id' is not validate: id = {}",
+                    id);
+        }
     }
 
     @Override
@@ -87,7 +182,19 @@ public class UserDiscountServiceImpl implements UserDiscountService{
         LOG.debug("deleteUserDiscountByDiscount: discount_id = {}",
                 discountId);
 
-        userDiscountMapper.deleteUserDiscountByDiscount(discountId);
+        if (discountId != null && discountId >= 0){
+            DiscountDTO discountDTO = discountService.getDiscount(discountId);
+            if (discountDTO != null) {
+                userDiscountMapper.deleteUserDiscountByDiscount(discountId);
+            }
+            else {
+                LOG.debug("Error in 'deleteUserDiscountByDiscount'! Discount is not found");
+            }
+        }
+        else {
+            LOG.debug("Error in 'deleteUserDiscountByDiscount'! 'id' is not validate: id = {}",
+                    discountId);
+        }
     }
 
     @Override
@@ -95,6 +202,18 @@ public class UserDiscountServiceImpl implements UserDiscountService{
         LOG.debug("deleteUserDiscountByUser: user_id = {}",
                 userId);
 
-        userDiscountMapper.deleteUserDiscountByUser(userId);
+        if (userId != null && userId >= 0){
+            UserDTO userDTO = userService.getUser(userId);
+            if (userDTO != null) {
+                userDiscountMapper.deleteUserDiscountByUser(userId);
+            }
+            else {
+                LOG.debug("Error in 'deleteUserDiscountByUser'! User is not found");
+            }
+        }
+        else {
+            LOG.debug("Error in 'deleteUserDiscountByUser'! 'id' is not validate: id = {}",
+                    userId);
+        }
     }
 }
